Fix URL normalization
For a line like
    deb http://apt.knut.univention.de/ ucs500 main
apt-mirror uses sanitise_uri($uri), which returns the empty string "" in
that case. This is used as a path and concatenated with "/dist/...",
which in an *absolute* path instead of a relative path relative to the
current working directory $skel.

index = dists/ucs500/main/binary-amd64/Packages
path =
mirror = /var/lib/univention-repository/mirror/

> apt-mirror: can't open index /dists/ucs500/main/binary-amd64/Packages in process_index at /usr/bin/apt-mirror line 796, <STREAM> chunk 144.

Gbp-Pq: Topic bugfix